.cursor/extensions/colony-os/requirements.txt
# Colony OS Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
supabase>=2.0.0

